Add files via upload
A face mask detection project is a computer vision system designed to identify whether a person is wearing a face mask or not. The system uses machine learning algorithms to analyze images or video feeds in real-time, identifying the presence or absence of a face mask on each person within the frame.

This project has gained popularity due to the COVID-19 pandemic, as face masks have become a common way to prevent the spread of the virus. The system can be integrated into various settings, such as public spaces, transportation systems, schools, hospitals, and offices to ensure that individuals are complying with the face mask mandate.The project typically involves several stages, including data collection, data preprocessing, model training, and model deployment. Data collection involves gathering images and videos of people wearing and not wearing face masks. Data preprocessing involves cleaning and organizing the data to ensure it is suitable for model training. Model training involves using machine learning algorithms to create a predictive model that can accurately detect the presence or absence of a face mask in an image or video feed. Finally, model deployment involves integrating the trained model into a system that can be used in real-time settings.

Overall, a face mask detection project is an important tool for ensuring compliance with face mask mandates and helping to prevent the spread of COVID-19 in public spaces.